+++ angular-lib/projects/fittable-angular/src/lib/controls-view/cell-editor/cell-editor.component.ts
@@ -22,7 +22,7 @@ import {
 @Component({
   selector: 'fit-cell-editor',
   templateUrl: './cell-editor.component.html',
-  styleUrls: ['./cell-editor.component.css'],
+  styleUrls: ['../../common/css/fittable-main.css'],
 })
 export class CellEditorComponent implements OnInit, AfterViewInit, OnDestroy {
   @Input() cellEditorListener!: CellEditorListener;