+++ angular-lib/projects/fittable-angular/src/lib/controls-view/combo/combo.component.ts
@@ -16,7 +16,7 @@ import { createToggleStyle } from '../common/style-functions.model';
 @Component({
   selector: 'fit-combo',
   templateUrl: './combo.component.html',
-  styleUrls: ['../common/css/controls-common.css', './combo.component.css'],
+  styleUrls: ['../../common/css/fittable-main.css'],
 })
 export class ComboComponent extends PopupControlComponent implements OnInit {
   @Input() override model!: PopupControl;