+++ angular-lib/projects/fittable-angular/src/lib/controls-view/popup-button/popup-button.component.ts
@@ -14,10 +14,7 @@ import { createToggleStyle } from '../common/style-functions.model';
 @Component({
   selector: 'fit-popup-button',
   templateUrl: './popup-button.component.html',
-  styleUrls: [
-    '../common/css/controls-common.css',
-    './popup-button.component.css',
-  ],
+  styleUrls: ['../../common/css/fittable-main.css'],
 })
 export class PopupButtonComponent
   extends PopupControlComponent