+++ angular-lib/projects/fittable-angular/src/lib/controls-view/settings-bar/settings-button/settings-button.component.ts
@@ -10,10 +10,7 @@ import { PopupControlComponent } from '../../common/popup-control-component.mode
 @Component({
   selector: 'fit-settings-button',
   templateUrl: './settings-button.component.html',
-  styleUrls: [
-    '../../common/css/controls-common.css',
-    './settings-button.component.css',
-  ],
+  styleUrls: ['../../../common/css/fittable-main.css'],
 })
 export class SettingsButtonComponent
   extends PopupControlComponent