+++ angular-lib/projects/fittable-angular/src/lib/controls-view/statusbar/statusbar.component.ts
@@ -4,22 +4,8 @@ import { Statusbar } from 'fittable-core/view-model';
 
 @Component({
   selector: 'fit-statusbar',
-  template: '<div class="statusbar">{{model.getText()}}</div>',
-  styles: [
-    `
-      .statusbar {
-        position: relative;
-        padding-right: 0.875rem;
-        border-top: 1px solid var(--statusbar-border-color);
-        background-color: var(--statusbar-background-color);
-        color: var(--statusbar-color);
-        width: auto;
-        text-align: right;
-        font-size: 0.625rem;
-        line-height: 1.2rem;
-      }
-    `,
-  ],
+  template: '<div class="fit-statusbar">{{model.getText()}}</div>',
+  styleUrls: ['../../common/css/fittable-main.css'],
 })
 export class StatusbarComponent implements OnDestroy {
   @Input() model!: Statusbar;